Add exercises 4.15-4.23

+++ part4/controllers/blogs.js
@@ -1,33 +1,47 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
-const logger = require('../utils/logger')
+const middleware = require('../utils/middleware')
+
+
 
 blogsRouter.get('/', async (request, response) => {
-  const blogs = await Blog.find({})
+  const blogs = await Blog
+    .find({}).populate('user', { username: 1, name: 1 })
+
   response.json(blogs)
 })
 
-blogsRouter.post('/', async (request, response) => {
-    const body = request.body
-    logger.info("req")
-    logger.info(request)
-    logger.info("body")
-    logger.info(body)
 
-    const blog = new Blog({
-      title: body.title,
-      author: body.author,
-      url: body.url,
-      likes: body.likes || 0
-    })
-    
-    const savedBlog = await blog.save()
-    response.status(201).json(savedBlog)
+blogsRouter.post('/', middleware.userExtractor, async (request, response) => {
+  const body = request.body
+
+  const user = request.user
+
+  const blog = new Blog({
+    title: body.title,
+    author: body.author,
+    url: body.url,
+    user: user.id,
+    likes: body.likes || 0
+  })
+
+  const savedBlog = await blog.save()
+  user.blogs = user.blogs.concat(savedBlog._id)
+  await user.save()
+
+  response.status(201).json(savedBlog)
 })
 
-blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndDelete(request.params.id)
-  response.status(204).end()
+blogsRouter.delete('/:id', middleware.userExtractor, async (request, response) => {
+  const user = request.user
+  const blog = await Blog.findById(request.params.id)
+
+  if (user && blog && user._id.toString() === blog.user._id.toString() ) {
+    await Blog.findByIdAndDelete(request.params.id)
+    response.status(204).end()
+  }
+
+
 })
 
 blogsRouter.put('/:id', async (request, response) => {
